feat: remove dotation filter from simulation detail
