+++ gsl_projet/views.py
@@ -1,21 +1,12 @@
-from django.db.models import Q
-from django.forms import NumberInput, Select
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
 from django.views.decorators.http import require_GET
 from django.views.generic import ListView
-from django_filters import (
-    ChoiceFilter,
-    FilterSet,
-    MultipleChoiceFilter,
-    NumberFilter,
-)
 from django_filters.views import FilterView
 
-from gsl_demarches_simplifiees.models import Dossier
 from gsl_projet.services import ProjetService
-from gsl_projet.utils.django_filters_custom_widget import CustomCheckboxSelectMultiple
 from gsl_projet.utils.filter_utils import FilterUtils
+from gsl_projet.utils.projet_filters import ProjetFilters
 
 from .models import Projet
 
@@ -86,130 +77,7 @@ def get_projet(request, projet_id):
     return render(request, "gsl_projet/projet.html", context)
 
 
-class ProjetFilters(FilterSet):
-    dotation = ChoiceFilter(
-        field_name="dossier_ds__demande_dispositif_sollicite",
-        choices=Dossier.DEMANDE_DISPOSITIF_SOLLICITE_VALUES,
-        widget=Select(
-            attrs={
-                "class": "fr-select",
-                "onchange": "this.form.submit()",
-                "placeholder": "Toutes les dotations",
-            }
-        ),
-    )
-
-    porteur = ChoiceFilter(
-        field_name="dossier_ds__porteur_de_projet_nature__label__in",
-        choices=(
-            ("EPCI", "EPCI"),
-            ("Communes", "Communes"),
-        ),
-        method="filter_porteur",
-        widget=Select(
-            attrs={
-                "class": "fr-select",
-                "onchange": "this.form.submit()",
-                "placeholder": "Tous les porteurs",
-            },
-        ),
-    )
-
-    def filter_porteur(self, queryset, _name, value):
-        return queryset.filter(
-            dossier_ds__porteur_de_projet_nature__label__in=ProjetService.PORTEUR_MAPPINGS.get(
-                value
-            )
-        )
-
-    cout_min = NumberFilter(
-        method="filter_cout_min",
-        widget=NumberInput(
-            attrs={"class": "fr-input", "min": "0"},
-        ),
-    )
-
-    def filter_cout_min(self, queryset, _name, value):
-        return queryset.filter(
-            Q(assiette__isnull=False, assiette__gte=value)
-            | Q(assiette__isnull=True, dossier_ds__finance_cout_total__gte=value)
-        )
-
-    cout_max = NumberFilter(
-        method="filter_cout_max",
-        widget=NumberInput(
-            attrs={"class": "fr-input", "min": "0"},
-        ),
-    )
-
-    def filter_cout_max(self, queryset, _name, value):
-        return queryset.filter(
-            Q(assiette__isnull=False, assiette__lte=value)
-            | Q(assiette__isnull=True, dossier_ds__finance_cout_total__lte=value)
-        )
-
-    montant_demande_max = NumberFilter(
-        field_name="dossier_ds__demande_montant",
-        lookup_expr="lte",
-        widget=NumberInput(
-            attrs={"class": "fr-input", "min": "0"},
-        ),
-    )
-
-    montant_demande_min = NumberFilter(
-        field_name="dossier_ds__demande_montant",
-        lookup_expr="gte",
-        widget=NumberInput(
-            attrs={"class": "fr-input", "min": "0"},
-        ),
-    )
-
-    montant_retenu_min = NumberFilter(
-        field_name="dossier_ds__annotations_montant_accorde",
-        lookup_expr="gte",
-        widget=NumberInput(
-            attrs={"class": "fr-input", "min": "0"},
-        ),
-    )
-
-    montant_retenu_max = NumberFilter(
-        field_name="dossier_ds__annotations_montant_accorde",
-        lookup_expr="lte",
-        widget=NumberInput(
-            attrs={"class": "fr-input", "min": "0"},
-        ),
-    )
-
-    status = MultipleChoiceFilter(
-        field_name="dossier_ds__ds_state",
-        choices=Dossier.DS_STATE_VALUES,
-        widget=CustomCheckboxSelectMultiple(),
-    )
-
-    class Meta:
-        model = Projet
-        fields = (
-            "dotation",
-            "porteur",
-            "cout_min",
-            "cout_max",
-            "montant_demande_min",
-            "montant_demande_max",
-            "montant_retenu_min",
-            "montant_retenu_max",
-            "status",
-        )
-
-    @property
-    def qs(self):
-        self.queryset = Projet.objects.all()
-        qs = super().qs
-        qs = ProjetService.add_ordering_to_projets_qs(qs, self.request.GET.get("tri"))
-        return qs
-
-
-# TODO : rename this class
-class ProjetFiltersFils(ProjetFilters):
+class ProjetListViewFilters(ProjetFilters):
     @property
     def qs(self):
         qs = super().qs
@@ -228,7 +96,7 @@ class ProjetFiltersFils(ProjetFilters):
 class ProjetListView(FilterView, ListView, FilterUtils):
     model = Projet
     paginate_by = 25
-    filterset_class = ProjetFiltersFils
+    filterset_class = ProjetListViewFilters
     template_name = "gsl_projet/projet_list.html"
 
     def get_context_data(self, **kwargs):
